just a little refactor to clear up some confusion.
Did you know that "foo.bar++" throws no error when bar is undefined?

Now we keep counts in a dedicated little object, and toss that around.

For now we construct a new dedicated assertions object for each test case
taht we run, and we try to keep the coupling loose between the two.

We will have more work to do when we deal with running a filetree, etc
of multiple files with tests in them.