fix exact match (literal string) testing
@to-document:{

== on using the --testcase flag

Internally when you use test() to add test methods
your test methods are stored in your test case object with
a name beginning with "test " prepended to the string you
use to name the test.  e.g. with:

  testcase.test('foo bar', function(){..});

the test is stored internally with a name "test for bar".

If you perfer you can add test methods explicity in
the following manner:

  testcase['test foo bar'] = function(){..});

which will have the exact same effect.

As a third option, you can add test methods by naming them
in camel case with the string 'test' beginning the name,
e.g.

  testcase.testFooBar = function(){..}

Which you use is a matter of personal taste or team convention.
(If you're curious, all test methods will be found using a regex
something like /^test[ 0-9A-Z]/).

When using the --testcase=NAME or --tescase=/REGEX/ flag,
the leading "test" or "test " string will *not* be used
when executing the search, so write the search pattern
as if the leading "test" string is not there, e.g.:

  --testcase='foo bar'   // for first example
or
  --testcase=/^foo/      // as regex anchored to beginning
or
  --testcase='FooBar'    // for the last variant.

}

this is all contingent on lines 273-275 in this commit