napsuite now lets test cases have child test cases
@to-document: {
Napsuite test cases can have child test cases, added with the
childCase() method.

All test cases are still effectively added to the 'global namespace',
that is, if you are filtering by testcase name, and that testcase name
does not match the --testcase regex for example, any children it has
will still be considered for running, and so on recursively.

This is a) experimental and b) by design.  It is thought that developers
will want to refer to test cases to run and not run just by their
raw basename only, and not some crazy fully qualified name that
reaches up to the parent.

Also, the uniqueness of testcase names is of course not enforced, and
since they have a flat global namespace, .. etc.
}