Finally !! we have syntax highlighting !!
I have to say i'm quite in love with how it turend out.

This looks like a lot of code! partly it's because we took care
to give helpful warning messages when users try to load invalid brushes
or themes.  We also provide 4 "configuration points" for themes:
1) specify them block-specifically, in the document metadata
2) specify them document wide in the document metadata (with '*')
3) they can be specified in build.js with "defaultTheme"
4) (hard coded:) they can appear in config.js and is hard coded a default theme.

We do yet more ugly hacks to dance around to be able to load
stylesheets into specific pages.  Maybe this can be refactored?