(decided we don't like the look of config/config.js)
it's too confusing to have a file named "config" in version control,
when really what this file is is "deps paths", that is, paths
to the dependencies we have.

So really, for now it shouldn't be that visible to the client.
You have to have these libraries with these (which? @todo) versions
in vendor/ with these filenames or it won't work!