(Change generated files: start using common file names for doc pages)
this is kind of a hack so we can keep regenerating our tree
and getting it to ignore our useful files that we have changed,
while still getting the changes in generated files (e.g. style.css)
that we want.  Before this, we would get crufty files we didn't want
each time we regenerated.

We regenerate because we don't want to version (or worry about
if it has changes) files that we have not chnaged from the
template.